Move UpdateError to updater.spi package
