+++ src/main/java/org/opentripplanner/updater/trip/metrics/BatchTripUpdateMetrics.java
@@ -11,9 +11,9 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
-import org.opentripplanner.model.UpdateError;
-import org.opentripplanner.model.UpdateSuccess;
+import org.opentripplanner.updater.spi.UpdateError;
 import org.opentripplanner.updater.spi.UpdateResult;
+import org.opentripplanner.updater.spi.UpdateSuccess;
 import org.opentripplanner.updater.trip.UrlUpdaterParameters;
 
 /**
